requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
Flask-JWT-Extended==4.4.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.3
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.1
six==1.16.0
SQLAlchemy==1.4.36
typing_extensions==4.2.0
Werkzeug==2.1.2
wrapt==1.14.1
